Add day tracking and daily item respawn to GameTimeUI
Introduced a day counter and date UI to the BeachScene and GameTimeUI, updating the UI to display the current day. On day change, all ItemSpawner instances in the scene are triggered to respawn items. Also made SpawnItemsOnStart public in ItemSpawner, adjusted player speed, and updated various asset and font settings to support the new UI.